+++ decomposition/Sub_Matrix_Decomposition_Cost_Function.cpp
@@ -40,15 +40,15 @@ double get_submatrix_cost_function(Matrix& matrix) {
     // ********************************
 
     // number of submatrices
-    size_t submatrices_num = 4;
+    int submatrices_num = 4;
 
-    size_t submatrices_num_row = 2;
+    int submatrices_num_row = 2;
 
 
     // Extracting submatrices from the unitary
     std::vector<Matrix, tbb::cache_aligned_allocator<Matrix>> submatrices(submatrices_num);
 
-    tbb::parallel_for( tbb::blocked_range<size_t>(0, submatrices_num, 1), functor_extract_submatrices( matrix, &submatrices ));
+    tbb::parallel_for( tbb::blocked_range<int>(0, submatrices_num, 1), functor_extract_submatrices( matrix, &submatrices ));
 
 
     
@@ -58,7 +58,7 @@ double get_submatrix_cost_function(Matrix& matrix) {
     // ********************************
 
 
-    size_t prod_num = submatrices_num*submatrices_num_row;
+    int prod_num = submatrices_num*submatrices_num_row;
     tbb::combinable<double> priv_prod_cost_functions{[](){return 0;}};
 
     tbb::parallel_for(0, (int) prod_num, 1, functor_submtx_cost_fnc( &submatrices, &priv_prod_cost_functions, prod_num ));
@@ -97,27 +97,27 @@ functor_extract_submatrices::functor_extract_submatrices( Matrix& matrix_in, std
 @brief Operator to extract the submatrix indexed by submtx_idx
 @param r A range of indices labeling the given submatrix to be extracted
 */
-void functor_extract_submatrices::operator()( tbb::blocked_range<size_t> r ) const {
+void functor_extract_submatrices::operator()( tbb::blocked_range<int> r ) const {
   
     // number of submatrices
-    size_t submatrices_num_row = 2;
+    int submatrices_num_row = 2;
 
     // number of columns in the input matrix
-    size_t matrix_size = matrix.rows;
+    int matrix_size = matrix.rows;
 
     // number of columns in the submatrices
-    size_t submatrix_size = matrix_size/2;
+    int submatrix_size = matrix_size/2;
 
-    for ( size_t submtx_idx = r.begin(); submtx_idx != r.end(); submtx_idx++) {
+    for ( int submtx_idx = r.begin(); submtx_idx != r.end(); submtx_idx++) {
 
         // The given Submatrix
         Matrix& submatrix = (*submatrices)[ submtx_idx ];
 
 
         // create submarix data by striding the original matrix
-        size_t jdx = submtx_idx % submatrices_num_row;
-        size_t idx = (size_t) (submtx_idx-jdx)/submatrices_num_row;
-        size_t matrix_offset = idx*(matrix.stride*submatrix_size) + jdx*(submatrix_size);
+        int jdx = submtx_idx % submatrices_num_row;
+        int idx = (int) (submtx_idx-jdx)/submatrices_num_row;
+        int matrix_offset = idx*(matrix.stride*submatrix_size) + jdx*(submatrix_size);
         submatrix = Matrix(matrix.get_data()+matrix_offset, submatrix_size, submatrix_size, matrix.stride);
 
 
@@ -127,14 +127,14 @@ void functor_extract_submatrices::operator()( tbb::blocked_range<size_t> r ) con
 
 
         // extract the submatrix
-        size_t jdx = submtx_idx % submatrices_num_row;
-        size_t idx = (size_t) (submtx_idx-jdx)/submatrices_num_row;
+        int jdx = submtx_idx % submatrices_num_row;
+        int idx = (int) (submtx_idx-jdx)/submatrices_num_row;
 
         // copy memory to submatrices
-        for ( size_t row_idx=0; row_idx<submatrix_size; row_idx++ ) {
+        for ( int row_idx=0; row_idx<submatrix_size; row_idx++ ) {
 
-            size_t matrix_offset = idx*(matrix_size*submatrix_size) + jdx*(submatrix_size) + row_idx*matrix_size;
-            size_t submatrix_offset = row_idx*submatrix_size;
+            int matrix_offset = idx*(matrix_size*submatrix_size) + jdx*(submatrix_size) + row_idx*matrix_size;
+            int submatrix_offset = row_idx*submatrix_size;
             memcpy(submatrix.get_data()+submatrix_offset, matrix.get_data()+matrix_offset, submatrix_size*sizeof(QGD_Complex16));
 
         }
@@ -161,7 +161,7 @@ void functor_extract_submatrices::operator()( tbb::blocked_range<size_t> r ) con
 @param prod_num_in The number of partial cost function values (equal to the number of distinct submatrix products.)
 @return Returns with the instance of the class.
 */
-functor_submtx_cost_fnc::functor_submtx_cost_fnc( std::vector<Matrix, tbb::cache_aligned_allocator<Matrix>>* submatrices_in, tbb::combinable<double>* prod_cost_functions_in, size_t prod_num_in ) {
+functor_submtx_cost_fnc::functor_submtx_cost_fnc( std::vector<Matrix, tbb::cache_aligned_allocator<Matrix>>* submatrices_in, tbb::combinable<double>* prod_cost_functions_in, int prod_num_in ) {
 
     submatrices = submatrices_in;
     prod_cost_functions = prod_cost_functions_in;
@@ -175,11 +175,11 @@ functor_submtx_cost_fnc::functor_submtx_cost_fnc( std::vector<Matrix, tbb::cache
 void functor_submtx_cost_fnc::operator()( int product_idx ) const {
 
     // number of submatrices
-    size_t submatrices_num_row = 2;
+    int submatrices_num_row = 2;
 
     // select the given submatrices used to calculate the partial cost_function
-    size_t jdx = product_idx % submatrices_num_row;
-    size_t idx = (size_t) ( product_idx - jdx )/submatrices_num_row;
+    int jdx = product_idx % submatrices_num_row;
+    int idx = (int) ( product_idx - jdx )/submatrices_num_row;
 
     // calculate the submatrix product
     Matrix tmp = (*submatrices)[jdx];
@@ -201,15 +201,15 @@ void functor_submtx_cost_fnc::operator()( int product_idx ) const {
 //submatrix_prod.print_matrix();
 
     // number of elements in the matrix of submatrix products
-    size_t submatrix_size = submatrix_prod.rows;
-    size_t element_num = submatrix_size*submatrix_size;
+    int submatrix_size = submatrix_prod.rows;
+    int element_num = submatrix_size*submatrix_size;
 
     // subtract the corner element from the diagonal
     QGD_Complex16 corner_element = submatrix_prod[0];
-    //tbb::parallel_for( tbb::blocked_range<size_t>(0, submatrix_size, 1), [&](tbb::blocked_range<size_t> r){
-//        for ( size_t row_idx=r.begin(); row_idx != r.end(); row_idx++) {
-        for ( size_t row_idx=0; row_idx < submatrix_size; row_idx++) {
-            size_t element_idx = row_idx*submatrix_prod.stride+row_idx;
+    //tbb::parallel_for( tbb::blocked_range<int>(0, submatrix_size, 1), [&](tbb::blocked_range<int> r){
+//        for ( int row_idx=r.begin(); row_idx != r.end(); row_idx++) {
+        for ( int row_idx=0; row_idx < submatrix_size; row_idx++) {
+            int element_idx = row_idx*submatrix_prod.stride+row_idx;
             submatrix_prod[element_idx].real = submatrix_prod[element_idx].real  - corner_element.real;
             submatrix_prod[element_idx].imag = submatrix_prod[element_idx].imag  - corner_element.imag;
         }
@@ -217,9 +217,9 @@ void functor_submtx_cost_fnc::operator()( int product_idx ) const {
     //});
 
     // Calculate the |x|^2 value of the elements of the submatrixproducts and add to the partial cost function
-    //tbb::parallel_for( tbb::blocked_range<size_t>(0, element_num, 1), [&](tbb::blocked_range<size_t> r){
-//        for (size_t idx = r.begin(); idx != r.end(); idx ++) {
-        for (size_t idx = 0; idx < element_num; idx++) {
+    //tbb::parallel_for( tbb::blocked_range<int>(0, element_num, 1), [&](tbb::blocked_range<int> r){
+//        for (int idx = r.begin(); idx != r.end(); idx ++) {
+        for (int idx = 0; idx < element_num; idx++) {
 
             // store the calculated value for the given submatrix product element
             double &prod_cost_function_priv = prod_cost_functions->local();
